qapi: Fix indent level on doc comments in json files
The texinfo doc generation doesn't care much about
indentation levels, but we would like to add a rST
backend, and rST does care about indentation.

Make the doc comments more strongly consistent about indentation
for multiline constructs like:

@arg: description line 1
      description line 2

Returns: line one
         line 2

so that there is always exactly one space after the
colon, and subsequent lines align with the first.

This commit is a purely whitespace change, and it does
not alter the generated .texi files (because the texi
generation code strips away all the extra whitespace).
This does mean that we end up with some over-length lines.

Note that when the documentation for an argument fits
on a single line like this:

@arg: one line only

then stray extra spaces after the ':' don't affect the
rST output, so I have not attempted to methodically
fix them, though the preference is a single space here too.

Signed-off-by: Peter Maydell <peter.maydell@linaro.org>
Message-Id: <20200206173040.17337-10-peter.maydell@linaro.org>